Add files via upload
The enclosed report explores a recent data mining project I conducted within the KNIME platform. The project was designed to use four different prediction models (two number-based and two set-based) to discover the determinants of the customer churning, which is the percentage of customers who stop using the company during a particular timeframe.